Function to return many sets' data using brickset.com search API

Function to return many sets' data using brickset.com search API

brickset_get_sets(auth_args = list(key = brickset_key(), username =
  brickset_username(), password = brickset_password(), userHash =
  brickset_hash(), authtime = brickset_authtime()), ...)

Arguments

auth_args
list containing key, username, password, userHash (if known), and authtime (if known). If list is full of NULL values (default), they will be substituted from the environment. Environment variables should be named .brickset_key, .brickset_username, .brickset_password, .brickset_authtime, and .brickset_hash, though the last two should generally not be user-specified. Auth variables are only necessary if variables are not cached in the global environment; if the global variables exist, this argument can be ignored entirely.

...
Variables as defined in https://brickset.com/api/v2.asmx?op=getSets Any variables not provided will be substituted with default values from the documentation with the exception of pageSize, which has been increased.

Examples

# NOT RUN {
brickset_get_sets(theme = "Architecture", pageSize = "10")
# }
if (exists(".brickset_key") & exists(".brickset_hash")) {
  brickset_get_sets(theme = "Architecture", pageSize = "10")
}