Get subthemes released under a theme

Get subthemes released under a theme

brickset_get_subthemes(theme, auth_args = list(key = brickset_key(),
  username = brickset_username(), password = brickset_password(), userHash
  = brickset_hash(), authtime = brickset_authtime()))

Arguments

theme
Theme name

auth_args
list containing key, username, password, userHash (if known), and authtime (if known). If list is full of NULL values (default), they will be substituted from the environment. Environment variables should be named .brickset_key, .brickset_username, .brickset_password, .brickset_authtime, and .brickset_hash, though the last two should generally not be user-specified. Auth variables are only necessary if variables are not cached in the global environment; if the global variables exist, this argument can be ignored entirely.

Examples

# NOT RUN {
brickset_auth()
themes <- brickset_get_themes()
subthemes <- dplyr::filter(themes, subthemecount > 0) %>%
magrittr::extract2("theme") %>%
purrr::map_df(brickset_get_subthemes)
# }